+++ src/d3d12/d3d12-command.cpp
@@ -890,10 +890,10 @@ void CommandRecorder::cmdSetRenderState(const commands::SetRenderState& cmd)
         {
             const ScissorRect& src = state.scissorRects[i];
             D3D12_RECT& dst = scissorRects[i];
-            dst.left = LONG(src.minX);
-            dst.top = LONG(src.minY);
-            dst.right = LONG(src.maxX);
-            dst.bottom = LONG(src.maxY);
+            dst.left = src.minX;
+            dst.top = src.minY;
+            dst.right = src.maxX;
+            dst.bottom = src.maxY;
         }
         m_cmdList->RSSetScissorRects(state.scissorRectCount, scissorRects);
     }
